[vendor] Optimize IsAbsent deserialization
Boolean values are encoded as varint 0 or 1 in protobuf, and thus take
one byte in serialized form. This makes them fixed-length types, and we
can allocate all the memory we're going to need to deserialize repeated
fields of them in one go.

When we're not dealing with packed fields of IsAbsent booleans (for some
reason?), we can abuse our knowledge of the incoming data to allocate an
array of length Values for them.

This patch will get overwritten whenever we regenerate the protobuf
client code, and will need to be reapplied when we do, until we get this
upstream.