Monitor service uptime
API and Zipper expose an increasing counter 'Uptime' that counts
the minutes since each service was started. The idea behind it is that

    transformNull(nonNegativeDerivative(hostname.uptime))

will (once properly scaled) be equal to 1 when the service is fine, and
0 when it has just restarted. Averaging these across the fleet then
yields a metric between 0 and 1 that indicates whether the fleet is
generally happy with its life (near 1 is good) or whether systemd keeps
bringing it howling back into existance.

Changes in this metric during and after rollouts would indicate that the
new code is unstable, that this instability has inexplicably escaped our
attention during testing, and a revert is necessary.

We don't use uptime in seconds because the highest resolution of our
namespaces in our in-house Graphite setup is minutely (boo!). Using
seconds of uptime might thus give a weaker signal when something is
wrong.

In case this changes in the future, we can make the step size of the
counter configurable and log secondly, 10-secondly or whatever-sized
chunks of uptime. We thus don't encode the size of the uptime steps in
the metric name.