Monitor limiter saturation
Each server we talk to has a limiter that controls how many requests we
can send to it at once. This commit exports the ratio of used / total
space in each limiter as an expvar that can be inspected when debugging.

We also export the maximum of all those ratios as a Graphite metric, as
it's doubtful we want to involve the hostnames in the metric paths
there. A high maximum on a per-host basis, or across the fleet, is
perhaps a useful sign of bottlenecks.