Remove GraphTemplates
It's the only remaining part of the code that uses viper. Since we're by
now not compatible with graphite-web by not having Graph templates, and
we're also not compatible with graphite-web by having this
implementation of Graph templates, let's settle on the noncompatibility
that brings us the least pain for now.

See

    https://github.com/go-graphite/carbonapi/issues/163

for more.