Remove ignoreClientTimeout
It was added in a3bd7f1a with no indication of its purpose. My best
guess is that the idea behind it is that we may time out while serving a
heavy request, still finish it and populate our cache with the result,
and serve future requests for the same request from cache. I don't think
this is actually doing a lot for us though, because our timeouts are
so big that by the time this would happen, the results in the cache
would already be out of date.

The cfg package still parses ignoreClientTimeout, as the strict YAML
parsing would make parsing the existing (in production) configuration
files fail if we were to roll this out as is.