Cancel work in single- and multiGet
If a context times out, we cancel the requests that were in flight in
single- and multiGet.

This commit also obsoletes TimeoutAfterAllStarted from the config file.
We used to keep track of how many requests we were going to send, and
reset our timeout to TimeoutAfterAllStarted if we were able to send all
our requests before hitting our timeout.

That doesn't make sense to me. We already have timeouts that no living
being is waiting on, and this behaviour effectively doubles them. It
makes more sense to me to try to enforce the Global timeout we set. If
we want to be nice to queries that have spent a lot of time in our
limiter queues, we should go about that in a saner way.