RenderHandler: Unify render request making
We used to split the render request making into two cases depending on
whether we were sending globs as is (making a single request) or had
resolved them (and were thus making more requests). In this patch we
unify the two cases and make at least one request.

The metric fetching and computing pipeline used to be basically:

    for targets:
        resolve globs
        fetch metrics
        do math

In this series of patches, we're trying to work our way towards being
able to cancel work that has timed out. The naive way of doing that is:

    for targets:
        bail out if we timed out

        resolve globs
        fetch metrics
        do math

Manual testing has shown that this can still result in us taking
arbitrarily long to reply to requests. To avoid this, we need to make
the whole pipeline use concurrency primitives in a more fundamental way,
so it becomes something like

    for targets:
        select {
        case <- ctx.Done:
            bail

        case globs <- resolve:
            metrics <- fetch(globs)

        case stuff <- metrics:
            do math
        }

so we can cancel work effectively when we time out. To be able to do
this, it's nice if we pretend we always resolve globs for the fetching
step.

A more primitive alternative, sure to make Go enthusiasts weep, is to
keep track of when we received a request, when it will time out, and
have the pipeline be

    for targets:
        manual check for timeout
        resolve globs

        manual check for timeout
        fetch metrics

        manual check for timeout
        do math

It's more work on our part, but involves no concurrency primitives and
no major changes to the pipeline. Note that we'll need to keep track of
per-request timings like this anyway in the end to be able to
communicate timeout budgets to Zipper or any stores we have.

Either way, not having this case splitting is pretty nice on its own.