More detailed error messages in multiGet
We can stare at the causes of the errors we get in multiGet and bucket
them together by origin. Before we would implicitly assume that all
errors were due to hosts timing out. Making that more precise involves
making a surprising number of type switches on errors from the net
family of packages from the standard library, which we can hopefully
count on not changing radically very often.