Sort metrics by resolution before merging
Suppose we get three responses back for a metric of resolutions

    (A, B, A)

in that order, where A is a higher resolution than B. Previously, we'd
run through A to see if we could heal any given point, and if we saw one
we'd notice that B had different resolution, and then abandon the work.

If we instead sort the responses by resolution to get

    (A, A, B)

we maximize our chances of healing the metric.

This fixes the 4 breaking tests we wrote earlier.