Measure metric corruption
We define the ratio of corruption of a metric as the number of points we
healed in the metric, divided by the amount of points in the metric. We
also keep a running total of corruption that we export to Graphite.

We no longer log 'unable to merge ovalues'. We saw a lot of those
messages and they weren't actionable in any way.