Write 4 tests that break mergeResponses
We give mergeResponses three metrics to merge. One of them has a lower
resolution than the other two, and should be ignored, while the other
two are:

    (0, 1)    and    (1, 0)

In each case, the 0 value is 'missing'.

We pass the 3! = 6 permutations of the metric positions into
mergeRespones. In every case, the result of the merge should be

    (1, 1)

with no values missing. In 4 of the cases we setup, this fails.