Move internal endpoints to their own server
Making good on the threat made in 20f3f63d, we move our internal
endpoint to their own HTTP server that listens on a different port from
the main one.

This lets us pretend to users that we have a mostly static API, while we
experiment with interal controls somewhere that won't bother them.