Use loose config parsing by default
I've gotten myself into such a blind that I can't rollout successfully
without relaxing the strict config parsing.

Earlier, when I created the second HTTP server for our internal
endpoints, I gave it a default port that's the usual default port minus
1000. So far, so good.

When I then tried to replay some production traffic on a staging host
against the new setup, all my requests responded successfully within 50
ms. That is (a) amazing for this thing, and (b) so unbelievable that it
can't be right.

As the log files explained, the reason for this sudden spiffyness was
that:

(1) Carbonapi would start, and claim the default port for the new
    internal endpoints.

(2) Carbonzipper would then start, be denied that port, crash, and never
    come up again.

(3) Carbonapi would then spam prod traffic to a port no one was
    listening on, and because of our sycophantic approach to HTTP status
    codes, claimed that every one of those near-instant failures was a
    success.

The naive thing to do would now be to patch our Puppetry to set distinct
default internal ports for API and Zipper. If we did that, we'd bring
down all our production hosts over a span of 30 minutes, as one by one,
they'd be restarted by systemd, see an unknown configuration option, and
refuse to go in such an uncertain world.

Instead, let's silently ignore unknown configuration options, and hotfix
production with this commit before moving on with our lives.