Decode all RPC response types
We use a subset of the go-graphite/protocol V2 protobuf definition for
communication. We don't care about all the things defined in there, and
by defining a repeated Info response message we can avoid a nasty corner
case in api/zipper communication that breaks the /info handler in
current production.