Remove CarbonSearch
The #graphite team were the only ones using CarbonSearch, until a couple
of weeks ago. Now no one does, so let's remove it from the system for
now.

We have plans of coming back to this idea in the future, but until then
not having CarbonSearch makes the switch to pkg/backend easier.