Go back to protocol/carbonapi_v2_pb
The only reason I thought I had for having our own *.proto definitions
(that by doing so we could cleverly avoid the info response mangling)
was wrong, so let's use the same protobuf definitions as everyone else.

We'll avoid the info response mangling (which comes down to that we
don't know whether we're getting an InfoResponse or multiple
ServerInfoResponses) by staring at the byte's we're going to
deserialize, count the Protocol Buffer tags in there, and decide based
on that.