Actually try to find '*'
This took me the longest time to figure out, but if we tried to call

    /metrics/find?query=*

we'd get an internal server error, _every time_. Turns out that the
backend filter couldn't match '*' against any of the top-level domains
our clusters had and then assumed that none of them contained what we
were looking for.

We can guard against this in two ways by assuming that anything that
contains any of '*{}[]' matches any TLD.

Somewhat interestingly, _some_ of our backends must have been matching
this TLD, because we already guard against filtering out all the
backends.