Lay ground for using other protocols over HTTP
The HTTP backend currently only talks to go-carbon, which can serialize
responses in more than one format. We'd like to be able to talk those
formats, so this patch adds the option to use different deserializers to
unpack responses based on the Content-Type header of the response.

If we add support for deserializing pickle or msgpack formats, on which
some thoughts are documented in doc/wire-formats.adoc, we _should_ be
able to talk to Python carbon backends as well. If this happens, I think
we'll end up declaring in our config file what protocol a backend uses.
That seems more reliable than trying to figure that out on the fly.

The return tuple of `call` is as busy as we should allow it to become
now. If we decide we want more information about the response further
up, we should either define a special return type, or maybe just return
the *http.Response.