Trace execution of Requests
We keep track of the aggregated time in nanoseconds each request spends:

(1) Marshaling the request
(2) Waiting in limiters
(3) Making HTTP calls
(4) Reading the HTTP response body
(5) Unmarshaling responses

This lets us eyeball where we're spending our time on a request-level,
which is sometimes more helpful than using the CPU profiling tools.

We really should replace all of this with some form of tracing.