Backends with empty path cache contain nothing
This commit is deceptively short for what it does, and I don't know a
good way to write a regression test for what it ensures.

Consider the following scenario: We have two backends A and B, where A
is healthy but B responds with 404s to every request. We now send a find
request for '*' to them.

Our backends say they don't contain globs, so A will say it doesn't
contain '*'. However, if an empty path cache counts as containing
anything, then B will say it contains '*'. We'll thus send our request
to B _only_, which will return a 404.