Distinguish 4XX from 5XX errors in carbonzipper
A /render request that would return no data returns a 404 response.

A /metrics/find request that found nothing has to return a 200 response
with an empty body to make graphite-web 0.9.12 (which we still run in
production) happy with its life.

Both situations claim that a 404 client error happened to our
monitoring.

Some combination of this may make the reader unhappy, but such is life
in a land where we control none of the APIs we need to speak.