CSV output respects timezone parameter
The CSV output of carbonapi consists of lines of the form:

    "metric.name",YYYY-MM-DD hh:mm:ss,value

The time is the epoch timestamp of the metric datapoint expressed in the
local time zone.

When the "tz" option is set in the configuration file, or when a user
passes in a "tz" URL parameter, we should express the timestamp in that
time zone.