Be very careful about reading and closing response bodies
Since rolling out the pkg/backend refactor to Zipper, we've seen a lot
more open TCP connections on our frontends than before. Their numbers
grew from around 500 open connections at any time to around 50.000.

As far as I know, the most common (only?) source of TCP connection
leaking in the net/http library is not reading and closing response
bodies. In this patch, we are extra careful about always doing that when
we get a non-nil response in the first place.