Set idle connection timeout to 3 seconds
We see a lot of open idle connections on our production hosts. Our best
guess for why this happens is that a carbonzipper host will open a lot
of connections when it gets a busy request. Once that request has been
served, the now idle connections are never closed because of the default
value for http.Transport.IdleConnTimeout:

    https://golang.org/src/net/http/transport.go?s=3628:10127#L212

We don't know why so many connections are opened in the first place. The
number of connections we keep open has grown by a factor of 100 since we
rolled out the pkg/backend refactoring. This patch reverses that growth,
without bringing understanding into the underlying cause.