backend/net: terminate goroutine on canceled outbound request
The parent goroutine reads on the request channel, but stops doing so in
a select case on ctx.Done().  The goroutine that does the outbound
request may want to send the context.Canceled error into that abandoned
channel and block forever.

This change releases the goroutine doing the outbound request on context
cancellation by ensuring a successful send into a buffered channel.  It
is an alternate solution to #199.