lint: use golangci-lint, initial project config
I tried to create a meaningful project local config in .golangci.yml, but of
course this is a matter of taste and debate.  Most of the settings come from
the default golangci-lint config file.

As of now golangci-lint finds quite a bunch of issues.  IMHO they're all worth
investigating.  I have explicitly exluded "composite literal uses unkeyed
fields" in test files as this style is used widely in the table driven tests.

The make target "lint" shows only issues that are introduced since the
merge-base of this commit on the master branch in order to make the
introduction painless in CI pipelines etc.  Another target "lint-all" runs
golangci-lint w/o that restriction, i.e. it show "all the gore".

TODO: substitute staticcheck as the vendored-in version is old.

TODO: Set run.modules-download-mode=vendor in .golang.yml after integrating Go
modules. See #131, #190.

/fixes: #189