gonum: pin version and upgrade to 0.6.2
I'm trying to switch to go modules.

What we noticed in https://github.com/bookingcom/carbonapi/pull/190
is that after switching to go modules, this module gets updated.

In order to have a clean 'vendor' directory when switching to modules,
I'm first updating this library.

There are two functions using this library: linearRegression and polyfit.