Avereha/error (#275)
Logic change: when we have a render request and one query fails to parse, just answer with http 400 and return that error message. Do not answer with partial results. Before, we had the same behavior only if parsing failed during ParseExpr, but when parsing failed during getTargetData we would ignore this error.
I also tested how does graphite-web handle this and it seems to answer with the first http 500 for the parsing error that it finds.

    Answer with HTTP 400 for parsing errors during render requests. The error message will contain only the first parsing error encountered. Example response: Bad Request (400) Details: expression eval failed: missing argument "

    Cleanup mostly not used pkg/error dependency

    group all parsing error under ParserError type so we could check later if one error has this type or not.
    since I started using errors.As for this check, I removed pkg/errors dependency because it was not used anyway(we would wrap errors sometime, but we would not use that wrapping for anything) and I also wanted to avoid the mix of Cause and As/Is.

I think the logic change that I mentioned earlier is the best approach because that way we have consistent HTTP response for parsing errors.