Fixes for cairo consolidation
First fix: instead of consolidationg the consolidation, always
consolidate the original data.

Second fix: When there is no consolidation, we still need to set
ValuesPerPoint to 1. It it used later for computing the x coordinate,
and if we leave to 0 all x coordinated will be NaN(divison by zero).

I'm not very happy with this approach, a better one would be to have a
constructor for TimeSeries and set there the default value for
ValuesPerPoint to 1 there.