Avereha/apply by node (#302)
What issue is this change attempting to solve?

#146

Rewrite applyByNode so it work even if it's not the top level function.
For example, this expression was failing before:
highestCurrent(applyByNode(a.b.c,1,"%.d"),5)
How does this change solve the problem? Why is this the best approach?

I removed the whole 'RewriteFunction' functionality.

Before:

 for _, target := range(targets) {
   ParseExpr()
   GetTargetData()

GetTargetData() {
    Rewrite expr and append new expr to (targets)
    EvalExpr
}

After:

for _, target := range(targets) {
   ParseExpr()
   GetTargetData()
   EvalExpr(GetTargetData)
}
...

ApplyByNode(.., GetTargetData) {
  RewriteExpression()
  ParseNewExpression()
  GetTargetData() -> for new metrics that we just parsed
  EvalExpr()
}

I had to update signature for all functions to add two new parameters:

    GetTargetData -> function that retrieved new metric data for a parsed expression
    Context -> since I was updating all the functions signature I though to add the context too.

One alternative implementation that I considered was to put (parser.Expr, from, until, values) in one structure called EvaluationContext. That way, when we want to add new parameters to all the functions we just add a field in that struct instead of editing 90 files.
The reason I did not continue with this was that the PR was becoming really big and it would be very hard to review.
How can we be sure this works as expected?

make test.
Tested with production traffic.