renderHandler: check if context.Done()
During http request, we check if ctx.Done() and return with Err when
this happens.

But then when we collect results(and errors) we handle those context
deadline errors as regular errors.
For a complex request with 200k metrics,this is what happens:
 - ~120k metrics are retrieved with sucess
 - for the rest 80k metrics we received 'context deadline' errors
 When merging those errors, we just drop the 'context deadline' errors
 and continue operations with the remaining 120k metrics as if nothing
 happened.

With this check, we just return ctx.Err() on ctx.Done()