WriteTimeout & ctx.Deadline race condition
When processing a request, we use a ctx with a timeout that was
Timeout.Global value.

At the same time, we set the same Timeout as WriteTimeout when creating
the httpServer.

What happens it that after `Timeout.Global` seconds:
 - we check if ctx.Done() and return from functions with errors.
 - the http server is closing the connection because we reached
   WriteTimeout

So when we try to answer with an error, the connection will be closed.