avereha/prioritylimiter (#328)
* Add priority limiter

From time to time, simple requests are slow.

Extra logging helped us understand that the slowdown would happen when we would get a complex request to fill up all limiter slots in carbon API.
How does this change solve the problem? Why is this the best approach?

To prevent this issue, I'm using a priority queue for the limiter, with the priority set to the number of subrequests - fewer subrequests -> more priority.
That way, all the simple requests have the highest priority(0) in the carbon API.
Simple requests will be faster, complex requests will be slower.
For requests with equal priority, we are sorting them by uuid. That way, we can lower the number of active requests.

Possible future improvements:

    now that we know how many requests are waiting for the limiter and request size, we can start rejecting complex requests when we have more than X pending requests.
    we can propagate the priority in a header to the zipper

* carbonapi: add metrics for number of active/waiting requests

* carbonapi: register prometheus metrics