parser: return errors if there are syntax bugs in metric names wrapped by a function
In the current implementation, carbonapi returns 400 errors for queries like ns1.ns2.foo-[a b] or ns1.ns2.foo-{a, b}.
However, if these types of queries happened to be wrapped around within a function, like sumSeries, that accepts
metrics list, no errors are returned. As they are treated as individual metrics and sent to storage, storage would
returns 400 as well.

This commit fixes the issue, by generating a parser error if spaces and comma appears inappropriately within []/{}.
At the same time, nested range list (like [[]]) and unclosed ([/{) or unopened (]/}) ranges or value list are also
reported as errors, instead of treating them as proper query and forwarding them to storages.

As consequence of this change, some of the requests that are originally returning 200 would be returned with 400
status codes.