Add mismatch report offset config
Majorly this feature is developed because of rolled up metrics.
Data points are aggregated and rolled up in each of stores
individually and asynchronously. Therefore, it is very likely that
latest data points in the lower archive would have different
partial values. The new config adds the offset so the late
mismatches are not reported.
This helps the users to have a more clear objective and view on
the consistency of the data stored and returned.