Add source clusters to render logs
This commit adds a SourceClusters to metrics to identify the
cluster the metric originated from. When merging the metrics from
multiple upstreams, the result metric contains all of the backends
involved for that metric in its SourceClusters. In the end,
carbonapi, logs the request with all of the source clusters who
contributed for gathering the metrics.